fix: while loop block
